refactor(provider-node): reuse one struct that implements all chain client traits
