+++ provider-node/src/lib.rs
@@ -9,13 +9,10 @@
 //! - Coordinating provider-initiated checkpoints
 
 pub mod agreement_coordinator;
-pub(crate) mod agreement_coordinator_subxt;
 pub mod api;
 pub mod auth;
 pub mod challenge_responder;
-pub(crate) mod challenge_responder_subxt;
 pub mod checkpoint_coordinator;
-pub(crate) mod checkpoint_coordinator_subxt;
 pub mod cli;
 pub mod command;
 pub mod error;
@@ -24,10 +21,10 @@ pub mod fs_index;
 pub mod mmr;
 pub mod replica_sync;
 pub mod replica_sync_coordinator;
-pub(crate) mod replica_sync_coordinator_subxt;
 pub mod s3_api;
 pub mod s3_index;
 pub mod storage;
+pub(crate) mod subxt_client;
 pub mod types;
 
 pub use agreement_coordinator::{
