+++ provider-node/src/subxt_client.rs
@@ -1,42 +1,74 @@
-//! Subxt-based production chain client for the replica sync coordinator.
+//! Subxt-based production chain client shared by all coordinators.
+//!
+//! A single [`SubxtChainClient`] holds one [`subxt::OnlineClient`] connection
+//! and one signing key, and implements every coordinator's chain-client trait
+//! (`AgreementChainClient`, `CheckpointChainClient`, `ReplicaSyncChainClient`,
+//! `ChallengeChainClient`). Coordinators still depend on the narrow trait they
+//! need, so per-trait mocks keep working; the production wiring just hands each
+//! one a clone of the same client (a cheap `OnlineClient`/`Keypair` clone that
+//! shares the underlying WebSocket connection).
 
+use crate::agreement_coordinator::AgreementChainClient;
+use crate::challenge_responder::{ChallengeChainClient, DetectedChallenge};
+use crate::checkpoint_coordinator::{CheckpointChainClient, CheckpointDuty};
 use crate::replica_sync_coordinator::{
     BucketSnapshot, ReplicaAgreementInfo, ReplicaSyncChainClient,
 };
 use crate::Error;
+use sp_core::crypto::Ss58Codec;
 use sp_core::H256;
 use storage_primitives::BucketId;
 use subxt::dynamic::Value;
 use subxt::ext::scale_value::value;
 
 /// Production implementation that talks to the chain via subxt.
-pub struct SubxtReplicaSyncChainClient {
+///
+/// Cloning is cheap: `OnlineClient` shares one connection behind an `Arc`, and
+/// the `Keypair` clone is a key copy. This lets a single instance be shared
+/// across every background coordinator.
+#[derive(Clone)]
+pub struct SubxtChainClient {
     api: subxt::OnlineClient<subxt::PolkadotConfig>,
     signer: subxt_signer::sr25519::Keypair,
 }
 
-impl SubxtReplicaSyncChainClient {
-    /// Connect to the chain and create a signer from the provider state's keypair.
-    pub async fn connect(
-        chain_ws_url: &str,
-        keypair: &sp_core::sr25519::Pair,
-    ) -> Result<Self, Error> {
-        use sp_core::Pair;
-
+impl SubxtChainClient {
+    /// Connect to the chain and create a signer from the seed URI.
+    ///
+    /// The signer is derived from `seed` (e.g. `//Alice` or a mnemonic), which
+    /// reproduces the provider's registered account — the identity every
+    /// on-chain action must be signed by.
+    pub async fn connect(chain_ws_url: &str, seed: &str) -> Result<Self, Error> {
         let api = subxt::OnlineClient::<subxt::PolkadotConfig>::from_url(chain_ws_url)
             .await
             .map_err(|e| Error::Internal(format!("Failed to connect to chain: {e}")))?;
 
-        let raw = keypair.to_raw_vec();
-        let secret_bytes: [u8; 32] = raw[..32]
-            .try_into()
-            .map_err(|_| Error::Internal("Invalid secret key length".to_string()))?;
-        let signer = subxt_signer::sr25519::Keypair::from_secret_key(secret_bytes)
+        let uri: subxt_signer::SecretUri = seed
+            .parse()
+            .map_err(|e| Error::Internal(format!("Invalid seed URI: {e}")))?;
+        let signer = subxt_signer::sr25519::Keypair::from_uri(&uri)
             .map_err(|e| Error::Internal(format!("Failed to create signer: {e}")))?;
 
-        tracing::info!("Replica sync coordinator connected to {}", chain_ws_url);
+        tracing::info!(
+            "Chain client connected to {} as {}",
+            chain_ws_url,
+            sp_core::crypto::AccountId32::from(signer.public_key().0).to_ss58check()
+        );
 
         Ok(Self { api, signer })
+    }
+
+    /// Get the current (latest) block number.
+    ///
+    /// Backs `get_current_block` on both the checkpoint and replica-sync traits.
+    async fn current_block(&self) -> Result<u64, Error> {
+        let block = self
+            .api
+            .blocks()
+            .at_latest()
+            .await
+            .map_err(|e| Error::Internal(format!("Failed to get latest block: {e}")))?;
+        Ok(block.number() as u64)
     }
 
     /// Convert a multiaddr string to an HTTP endpoint.
@@ -206,15 +238,227 @@ impl SubxtReplicaSyncChainClient {
 }
 
 #[async_trait::async_trait]
-impl ReplicaSyncChainClient for SubxtReplicaSyncChainClient {
-    async fn get_current_block(&self) -> Result<u64, Error> {
-        let block = self
+impl AgreementChainClient for SubxtChainClient {
+    async fn fetch_pending_requests(
+        &self,
+        provider_account: &[u8; 32],
+    ) -> Result<Vec<BucketId>, Error> {
+        let our_bytes = *provider_account;
+        {
+            // Iterate ALL AgreementRequests entries on chain.
+            // Storage layout: DoubleMap<Blake2_128Concat(BucketId), Blake2_128Concat(AccountId), Request>
+            // Key bytes: [16 pallet_hash][16 storage_hash][16 blake2_hash + 8 bucket_id][16 blake2_hash + 32 account]
+            // Total = 32 (prefix) + 24 (key1) + 48 (key2) = 104 bytes
+            let storage_query = subxt::dynamic::storage("StorageProvider", "AgreementRequests", ());
+            let storage = self
+                .api
+                .storage()
+                .at_latest()
+                .await
+                .map_err(|e| Error::Internal(format!("Failed to get storage: {e}")))?;
+
+            let mut entries = storage.iter(storage_query).await.map_err(|e| {
+                Error::Internal(format!("Failed to iterate agreement requests: {e}"))
+            })?;
+
+            let mut bucket_ids = Vec::new();
+            let mut entry_count = 0u32;
+
+            while let Some(result) = entries.next().await {
+                let entry = match result {
+                    Ok(e) => e,
+                    Err(e) => {
+                        tracing::warn!("Error reading agreement request entry: {}", e);
+                        continue;
+                    }
+                };
+
+                entry_count += 1;
+                let key_bytes = &entry.key_bytes;
+                let key_len = key_bytes.len();
+
+                // Expected key length: 32 (prefix) + 24 (key1) + 48 (key2) = 104
+                if key_len < 104 {
+                    tracing::warn!("Unexpected key length {} (expected 104), skipping", key_len);
+                    continue;
+                }
+
+                // Account bytes at offset 72 (32 prefix + 16 blake2 + 8 bucket + 16 blake2)
+                let account_bytes = &key_bytes[72..104];
+
+                if account_bytes != our_bytes.as_slice() {
+                    continue;
+                }
+
+                // Bucket ID at offset 48 (32 prefix + 16 blake2 hash)
+                let bucket_id = match key_bytes[48..56].try_into() {
+                    Ok(bytes) => u64::from_le_bytes(bytes),
+                    Err(_) => {
+                        tracing::warn!("Failed to parse bucket ID from key bytes, skipping");
+                        continue;
+                    }
+                };
+
+                tracing::info!(
+                    "Found pending agreement request for us: bucket {}",
+                    bucket_id
+                );
+                bucket_ids.push(bucket_id);
+            }
+
+            if entry_count > 0 {
+                tracing::info!(
+                    "Scanned {} agreement request entries, {} for us",
+                    entry_count,
+                    bucket_ids.len()
+                );
+            }
+
+            Ok(bucket_ids)
+        }
+    }
+
+    async fn accept_agreement(&self, bucket_id: BucketId) -> Result<(), Error> {
+        let tx = subxt::dynamic::tx(
+            "StorageProvider",
+            "accept_agreement",
+            vec![Value::u128(bucket_id as u128)],
+        );
+
+        let progress = self
             .api
-            .blocks()
+            .tx()
+            .sign_and_submit_then_watch_default(&tx, &self.signer)
+            .await
+            .map_err(|e| {
+                Error::Internal(format!(
+                    "Failed to submit accept_agreement for bucket {bucket_id}: {e}"
+                ))
+            })?;
+
+        progress.wait_for_finalized_success().await.map_err(|e| {
+            Error::Internal(format!(
+                "accept_agreement tx failed for bucket {bucket_id}: {e}"
+            ))
+        })?;
+
+        tracing::info!(
+            "Auto-accepted agreement for bucket {} (finalized)",
+            bucket_id
+        );
+        Ok(())
+    }
+}
+
+#[async_trait::async_trait]
+impl CheckpointChainClient for SubxtChainClient {
+    async fn get_current_block(&self) -> Result<u64, Error> {
+        self.current_block().await
+    }
+
+    async fn fetch_checkpoint_config(
+        &self,
+        bucket_id: BucketId,
+    ) -> Result<Option<(u32, u32)>, Error> {
+        use subxt::dynamic::At;
+
+        let config_query = subxt::dynamic::storage(
+            "StorageProvider",
+            "CheckpointConfigs",
+            vec![Value::u128(bucket_id as u128)],
+        );
+        let storage = self
+            .api
+            .storage()
             .at_latest()
             .await
-            .map_err(|e| Error::Internal(format!("Failed to get latest block: {e}")))?;
-        Ok(block.number() as u64)
+            .map_err(|e| Error::Internal(format!("Failed to get storage: {e}")))?;
+
+        match storage
+            .fetch(&config_query)
+            .await
+            .map_err(|e| Error::Internal(format!("Failed to fetch config: {e}")))?
+        {
+            Some(val) => {
+                let decoded = val
+                    .to_value()
+                    .map_err(|e| Error::Internal(format!("Failed to decode config: {e}")))?;
+                let interval = decoded
+                    .at("interval")
+                    .and_then(|v| v.as_u128())
+                    .unwrap_or(100) as u32;
+                let grace_period = decoded
+                    .at("grace_period")
+                    .and_then(|v| v.as_u128())
+                    .unwrap_or(20) as u32;
+                Ok(Some((interval, grace_period)))
+            }
+            None => Ok(None),
+        }
+    }
+
+    async fn submit_checkpoint(
+        &self,
+        duty: &CheckpointDuty,
+        signatures: Vec<(String, String)>,
+    ) -> Result<H256, Error> {
+        let bucket_id = duty.bucket_id;
+        let mmr_root = duty.mmr_root;
+        let start_seq = duty.start_seq;
+        let leaf_count = duty.leaf_count;
+        let window = duty.window;
+
+        // Build signature tuples for the extrinsic
+        let mut sig_values = Vec::with_capacity(signatures.len());
+        for (account, sig) in &signatures {
+            let account_id: sp_core::crypto::AccountId32 =
+                sp_core::crypto::Ss58Codec::from_ss58check(account).map_err(|e| {
+                    Error::Internal(format!("Invalid SS58 account '{account}': {e:?}"))
+                })?;
+            let account_bytes: [u8; 32] = account_id.into();
+
+            let sig_bytes = hex::decode(sig.trim_start_matches("0x"))
+                .map_err(|e| Error::Internal(format!("Invalid signature hex: {e}")))?;
+
+            sig_values.push(value!((
+                Value::from_bytes(account_bytes),
+                Sr25519(Value::from_bytes(sig_bytes))
+            )));
+        }
+
+        let tx = subxt::dynamic::tx(
+            "StorageProvider",
+            "provider_checkpoint",
+            vec![
+                Value::u128(bucket_id as u128),
+                Value::from_bytes(mmr_root.as_bytes()),
+                Value::u128(start_seq as u128),
+                Value::u128(leaf_count as u128),
+                Value::u128(window as u128),
+                Value::unnamed_composite(sig_values),
+            ],
+        );
+
+        let tx_progress = self
+            .api
+            .tx()
+            .sign_and_submit_then_watch_default(&tx, &self.signer)
+            .await
+            .map_err(|e| Error::Internal(format!("Failed to submit tx: {e}")))?;
+
+        let _events = tx_progress
+            .wait_for_finalized_success()
+            .await
+            .map_err(|e| Error::Internal(format!("Transaction failed: {e}")))?;
+
+        Ok(H256::zero())
+    }
+}
+
+#[async_trait::async_trait]
+impl ReplicaSyncChainClient for SubxtChainClient {
+    async fn get_current_block(&self) -> Result<u64, Error> {
+        self.current_block().await
     }
 
     async fn fetch_replica_agreements(
@@ -502,5 +746,118 @@ impl ReplicaSyncChainClient for SubxtReplicaSyncChainClient {
         );
 
         Ok((0, 0)) // Position 0, payment extracted from events in production
+    }
+}
+
+#[async_trait::async_trait]
+impl ChallengeChainClient for SubxtChainClient {
+    async fn poll_challenges(&self) -> Result<Vec<DetectedChallenge>, Error> {
+        let _storage = self
+            .api
+            .storage()
+            .at_latest()
+            .await
+            .map_err(|e| Error::Internal(format!("Failed to get storage: {e}")))?;
+
+        // TODO: Implement proper storage query for Challenges
+        // For now, return empty - challenges would be detected via events
+        Ok(vec![])
+    }
+
+    async fn submit_response(
+        &self,
+        challenge_id: (u32, u16),
+        chunk_data: Vec<u8>,
+        mmr_proof: storage_primitives::MmrProof,
+        chunk_proof: storage_primitives::MerkleProof,
+    ) -> Result<H256, Error> {
+        let challenge_id_val = value!({
+            deadline: challenge_id.0 as u128,
+            index: challenge_id.1 as u128
+        });
+
+        // Dynamic arrays built from iterators, then embedded in the macro
+        let peaks = Value::unnamed_composite(
+            mmr_proof
+                .peaks
+                .iter()
+                .map(|p| Value::from_bytes(p.as_bytes()))
+                .collect::<Vec<_>>(),
+        );
+        let mmr_siblings = Value::unnamed_composite(
+            mmr_proof
+                .leaf_proof
+                .siblings
+                .iter()
+                .map(|s| Value::from_bytes(s.as_bytes()))
+                .collect::<Vec<_>>(),
+        );
+        let mmr_path = Value::unnamed_composite(
+            mmr_proof
+                .leaf_proof
+                .path
+                .iter()
+                .map(|b| Value::bool(*b))
+                .collect::<Vec<_>>(),
+        );
+
+        let mmr_proof_val = value!({
+            peaks: peaks,
+            leaf: {
+                data_root: Value::from_bytes(mmr_proof.leaf.data_root.as_bytes()),
+                data_size: mmr_proof.leaf.data_size as u128,
+                total_size: mmr_proof.leaf.total_size as u128
+            },
+            leaf_proof: {
+                siblings: mmr_siblings,
+                path: mmr_path
+            }
+        });
+
+        let chunk_siblings = Value::unnamed_composite(
+            chunk_proof
+                .siblings
+                .iter()
+                .map(|s| Value::from_bytes(s.as_bytes()))
+                .collect::<Vec<_>>(),
+        );
+        let chunk_path = Value::unnamed_composite(
+            chunk_proof
+                .path
+                .iter()
+                .map(|b| Value::bool(*b))
+                .collect::<Vec<_>>(),
+        );
+
+        let chunk_proof_val = value!({
+            siblings: chunk_siblings,
+            path: chunk_path
+        });
+
+        let response_val = value!(Proof {
+            chunk_data: Value::from_bytes(&chunk_data),
+            mmr_proof: mmr_proof_val,
+            chunk_proof: chunk_proof_val
+        });
+
+        let tx = subxt::dynamic::tx(
+            "StorageProvider",
+            "respond_to_challenge",
+            vec![challenge_id_val, response_val],
+        );
+
+        let tx_progress = self
+            .api
+            .tx()
+            .sign_and_submit_then_watch_default(&tx, &self.signer)
+            .await
+            .map_err(|e| Error::Internal(format!("Failed to submit tx: {e}")))?;
+
+        let _events = tx_progress
+            .wait_for_finalized_success()
+            .await
+            .map_err(|e| Error::Internal(format!("Transaction failed: {e}")))?;
+
+        Ok(H256::zero())
     }
 }
